Fix mbarrier on ampere
